Simplify core package
