--- packages/core/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@rebass/core",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "scripts": {
-    "prepare": "babel src -d dist"
-  },
-  "license": "MIT",
-  "repository": "rebassjs/rebass",
-  "dependencies": {
-    "react": "^16.8.6",
-    "theme-ui": "^0.2.25"
-  }
-}
