+++ packages/docs/src/pages/notes.mdx
@@ -16,12 +16,11 @@
   - [x] Image
 - [x] Component variants
 - [ ] Rebass preset
-- [ ] New components
-  - [ ] Forms
-    - [ ] Label
-    - [ ] Input
-    - [ ] Select
-    - [ ] Textarea
+- [ ] @rebass/forms
+  - [ ] Label
+  - [ ] Input
+  - [ ] Select
+  - [ ] Textarea
 - [ ] Test with gatsby-plugin-theme-ui
 
 ---
